Added TA service, controller, enums & updated some entities

+++ backend/src/main/java/com/example/backend/repository/AssignmentRepo.java
@@ -3,5 +3,5 @@ package com.example.backend.repository;
 import com.example.backend.entity.Assignment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AssignmentRepo extends JpaRepository<Assignment,Integer> {
+public interface AssignmentRepo extends JpaRepository<Assignment,Long> {
 }
